Split receivers into file per class
