+++ lib/mocha/default_receiver.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Mocha
+  class DefaultReceiver
+    def initialize(mock)
+      @mock = mock
+    end
+
+    def mocks
+      [@mock]
+    end
+  end
+end
