+++ lib/mocha/mock.rb
@@ -5,7 +5,7 @@ require 'mocha/expectation'
 require 'mocha/expectation_list'
 require 'mocha/invocation'
 require 'mocha/names'
-require 'mocha/receivers'
+require 'mocha/default_receiver'
 require 'mocha/method_matcher'
 require 'mocha/parameters_matcher'
 require 'mocha/argument_iterator'
