+++ lib/mocha/mockery.rb
@@ -3,7 +3,8 @@
 require 'mocha/central'
 require 'mocha/mock'
 require 'mocha/names'
-require 'mocha/receivers'
+require 'mocha/object_receiver'
+require 'mocha/any_instance_receiver'
 require 'mocha/state_machine'
 require 'mocha/logger'
 require 'mocha/configuration'
