+++ test/unit/default_receiver_test.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require File.expand_path('../../test_helper', __FILE__)
-require 'mocha/receivers'
+require 'mocha/default_receiver'
 
 class DefaultReceiverTest < Mocha::TestCase
   include Mocha
